requirements.txt
streamlit
pandas
plotly
Pillow
openpyxl
requests
flask
flask-cors
python-dotenv